requirements.txt
Pillow==9.3.0
setuptools==63.2.0
